Revert "Merge branch 'main' into WNP_Reunion_StartupTask"
This reverts commit 5bf4d5e87ccc0b623abb9e34bd9352ca9429c38b, reversing
changes made to 11483cfcbd9a5a47a18a07627ca2f4b481511700.

+++ dev/PushNotifications/PushNotificationRegistrationToken.cpp
@@ -9,9 +9,9 @@
 
 namespace winrt::Microsoft::Windows::PushNotifications::implementation
 {
-    PushNotificationRegistrationToken::PushNotificationRegistrationToken(uint32_t const& cookie, winrt::Windows::ApplicationModel::Background::IBackgroundTaskRegistration const& taskRegistration): m_cookie(cookie), m_taskRegistration(taskRegistration) { }
+    PushNotificationRegistrationToken::PushNotificationRegistrationToken(uint64_t const& cookie, winrt::Windows::ApplicationModel::Background::IBackgroundTaskRegistration const& taskRegistration): m_cookie(cookie), m_taskRegistration(taskRegistration) { }
 
-    uint32_t PushNotificationRegistrationToken::Cookie()
+    uint64_t PushNotificationRegistrationToken::Cookie()
     {
         return m_cookie;
     }
